Take activation channel dim from metatype
